Renamed modules to append _bindings and make them private to the crate. Separate out type definitions to its own *defs module
